+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- StaffImprovements
+- SkullImprovements
+- CardChanges
+
+### Fixed
+
+- Fixed tranquil gain description
+
 ## [0.1.0] - 2024-05-25
 
 ### Added